image_viewer/requirements.txt
asgiref==3.11.1
Django==6.0.2
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
sqlparse==0.5.5
psycopg[binary]==3.2.10
dj-database-url==2.3.0